Create README.md
This is code a Tic Tac Toe game in Java. I learnt how to create a game of Tic Tac Toe using the awt and swing graphics library.I learnt how to create a graphical user interface (GUI) to display the Tic Tac Toe game, and check win or tie conditions using a 2D array. some features I added is a JButton for restarting the game every time there is a win or tie. Another feature is score keeping for playerX and playerO and updating the score after every game.